Move ResponseMessageAdaptor to TCK module as it's only used there
