+++ test-suite-http-server-tck-azure-function-http/src/test/java/io/micronaut/http/server/tck/azurehttpfunction/HttpResponseMessageAdapter.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.micronaut.azure.function.http;
+package io.micronaut.http.server.tck.azurehttpfunction;
 
 import com.microsoft.azure.functions.HttpResponseMessage;
 import io.micronaut.core.annotation.Internal;
